theater::client

Module tcp

TCP Connection for Theater

Provides a low-level TCP connection to a Theater server with simple send and receive methods that can be used with tokio::select!

Structs

TheaterConnection

A client connection to a Theater server