theater::utils

Function resolve_reference

pub async fn resolve_reference(
    reference: &str,
) -> Result<Vec<u8>, ReferenceError>

Resolve a reference to a byte array. A reference can be a file path, a store path, or a URL. The reference is resolved to a byte array.

A store path is a reference to a file in the store and is prefixed with store:// A store path can be either a hash or a path to a label:

store:///hash/

store:///

A URL is any reference starting with http:// or https://

Everything else is treated as a file path.